requirements.txt
django
pymysql
python-dotenv
pillow
djangorestframework
markdown
django-filter